Move private methods out of trainer, simplify interface
